feat(inferno): extract model registry and projector management into core/inferno
Moves Inferno-owned model registry, model-format handling, and projector logic into core/inferno/model_registry.py so inference lifecycle code is no longer centered in core/model_state.py.

Keeps activation sequencing in core for now while delegating model-store config, projector selection/status, and download helpers to the new Inferno-owned module without changing product-visible behavior.

Adds focused unit and API coverage for the extracted registry and projector flows.

Closes #295
Refs #264
