+++ core/inferno/__init__.py
@@ -1,7 +1,8 @@
 """Inferno -- Potato OS inference layer.
 
 Inferno owns the runtime-facing side of inference: backend proxying,
-model family classification, and adapter processes.  Everything between
+model family classification, model registry, settings normalization,
+projector management, and adapter processes.  Everything between
 "Potato says run this model" and "here is the OpenAI-compatible response"
 lives here.
 
@@ -13,6 +14,7 @@ Potato (caller) provides:
     - Model filename and optional source URL (for projector resolution)
     - POTATO_MODEL_PATH env var (for LiteRT adapter startup)
     - Hardware/device/OS inputs (memory, device class, runtime binaries)
+    - ModelStoreConfig with filesystem paths and product-level defaults
 
 Inferno (this package) provides:
     - BackendProxyError            exception for proxy failures
@@ -25,6 +27,12 @@ Inferno (this package) provides:
     - is_gemma4_filename           detect Gemma 4 model files
     - projector_repo_for_model     resolve HuggingFace projector repo
     - recommended_runtime_for_model  preferred runtime family for a model
+    - build_llama_server_args      pure function to build CLI args
+    - ModelStoreConfig             filesystem/policy config for registry ops
+    - Model registry functions     ensure/save/register/delete/update state
+    - Format handling              model_format_for_filename, validate_model_url
+    - Settings normalization       normalize_model_settings, build_model_capabilities
+    - Projector management         build_model_projector_status, download, candidates
     - litert_adapter               standalone FastAPI app (core.inferno.litert_adapter:app)
 
 Inferno does NOT import from core.model_state, core.runtime_state,
@@ -42,10 +50,40 @@ from .backend import (
 )
 from .launch_config import build_llama_server_args
 from .model_families import (
+    build_model_projector_status,
+    default_projector_candidates_for_model,
     is_gemma4_filename,
     is_qwen35_filename,
     projector_repo_for_model,
     recommended_runtime_for_model,
+)
+from .model_registry import (
+    DEFAULT_MODEL_CHAT_SETTINGS,
+    DEFAULT_MODEL_VISION_SETTINGS,
+    MODELS_STATE_VERSION,
+    VALID_MODEL_EXTENSIONS,
+    ModelSettingsValidationError,
+    ModelStoreConfig,
+    any_model_ready,
+    apply_model_chat_defaults,
+    build_model_capabilities,
+    delete_model,
+    describe_model_storage,
+    discover_local_model_filenames,
+    download_default_projector_for_model,
+    ensure_models_state,
+    get_model_by_id,
+    is_qwen35_a3b_filename,
+    model_file_path,
+    model_file_present,
+    model_format_for_filename,
+    model_supports_vision_filename,
+    normalize_model_settings,
+    register_model_url,
+    resolve_model_runtime_path,
+    save_models_state,
+    update_model_settings,
+    validate_model_url,
 )
 
 __all__ = [
@@ -53,11 +91,39 @@ __all__ = [
     "BackendResponse",
     "ChatCompletionRepository",
     "ChatRepositoryManager",
+    "DEFAULT_MODEL_CHAT_SETTINGS",
+    "DEFAULT_MODEL_VISION_SETTINGS",
     "FakeLlamaRepository",
     "LlamaCppRepository",
+    "MODELS_STATE_VERSION",
+    "ModelSettingsValidationError",
+    "ModelStoreConfig",
+    "VALID_MODEL_EXTENSIONS",
+    "any_model_ready",
+    "apply_model_chat_defaults",
     "build_llama_server_args",
+    "build_model_capabilities",
+    "build_model_projector_status",
+    "default_projector_candidates_for_model",
+    "delete_model",
+    "describe_model_storage",
+    "discover_local_model_filenames",
+    "download_default_projector_for_model",
+    "ensure_models_state",
+    "get_model_by_id",
     "is_gemma4_filename",
+    "is_qwen35_a3b_filename",
     "is_qwen35_filename",
+    "model_file_path",
+    "model_file_present",
+    "model_format_for_filename",
+    "model_supports_vision_filename",
+    "normalize_model_settings",
     "projector_repo_for_model",
     "recommended_runtime_for_model",
+    "register_model_url",
+    "resolve_model_runtime_path",
+    "save_models_state",
+    "update_model_settings",
+    "validate_model_url",
 ]
